Make Modal Form appear and disappear
The Modal Form now appears with the click of the "Add Book" button, and
disappears when the x icon is clicked, or when a click is made outside
of the Modal Form. I originally tried to create and style the form
through the JavaScript, however I found that I could leave the form in
the html, leave the styling in the CSS, and simply use event listeners
in JavaScript to show and hide the form. I left up the commented out
JavaScript code that creates and stylizes the Modal Form for this commit
only just to show my work.