Add toggle function to book read status
Now when user clicks on the read status icon of a book card, the read
status and the read status icon toggle. This was done by adding an
onclick attribute with a built in anonymous function to the img element
when creating the card after the form is submitted. Additionally, I made
it so that the text found in the book card also changes colour to match
the colour of the icon. Finally, I chan ged the class names of the book
cards so that each book card contains a unique class representing the
book number found in the myLibrary array.