Give functionality to x icon in book card
Added an onclick attribute function in each book card's x icon that
removes the book from the myLibrary array and from the UI when clicked.
I made it so that every book card x icon gets a unique class name, just
like the img elements. This time however, I named the classes like so:

remove-"book number"

and I used a regular expression to match the book number portion of the
class name in order to link it back to the matching object in the
myLibrary array.