Write up README file
Additionally, I made a change to the displayBooks function where the
classes given to the image and x-icon elements are not based on the
length of the myLibrary array, but based on the i variable of the for
loop. The reason I did this is because in the future, I will be adding
storage capabilities to this project, and the books that are already
stored in the array will be generated in the UI when the website opens,
which means the classes will have to be named appropriately.