Declare readToggle function in constructor
Before, the script was adding the readToggle method to a Book object
every time it was being created. This does not have to be repeated and
defeats the purpose of using an object constructor. I removed the method
declaration from the event listener, and declared this method in the
Book constructor.