Fix long title bug
I found that if the user would type a long continuous title with no
spaces, then the title would overflow off of the book card. I remedied
this by adding a style of overflow-wrap: break-word to the title div.
This adding style can be found in the displayBooks function in
script.js.