Add more extensive client side form verification
The form of this project now includes valid and invalid styling of the
form inputs, and messages telling the user the current input is invalid
and how to fix it. Here is what I did to complete this feature:

I adjusted the form dimensions to fit the error messages for each
required input (which I’ve added in the html). The error messages are
hidden, and are made visible from input event listeners. The CSS now
adds input:focus:invalid and input:valid pseudo-classes. Finally, in the
html, I added a novalidate attribute to the form element, which disables
browser validation and browser messages regarding form validation.
Now the form validation is done through the submit form event listener.